Add primitive name mangling
As a result of my horrible choice of not using separators for function call arguments, I am forced to use symbol literals for `new` semantics. That will most likely be changed in the future, to better align with function definitions.
On top of that, the lack of type inference means the name mangling is horribly primitive and does not do anything other than differentiate functions. No collision proofing or anything.