Add indexing assignment, pointer types, casting
As well as that, I've added the char type and char literal.
Side note, the parser requires changes, function call parsing is completely broken. This is not the first time I've encountered this, it is the reason why `new/constructor` semantics are done using symbols for it.
As a permanent solution, separators will be added for function call arguments. Should also fix identifier parsing.