Reimplement old funcall syntax + expression args
After many fights with the parser, it is finally over. The old syntax for function calls, `function_name arg1, arg2` is back, and it is improved. Prior to this, it was both broken, and accepting only atoms as arguments. Now, finally, it works properly, and accepts expressions as arguments. The parsing fight is now over. Codegen is back on track!