Add array support
This took just way too much time for something that seemed simple at
first. There's a huge lack of info on the web on this, so it's just
trial and error, and hope that LLVM's codegen doesn't segfault.
Yes, LLVM itself segfaults. And it takes away all the IR with it, so you
got literally no clue why it segfaults. But such is life. Trial and
error, that's how the world was built, and that's how it'll die.