Add a raytracer benchmark
This is mostly to test how far the language has come so far. Each thing
that gets implemented shows the lack of very important features. But
only by trial and error things will advance!

Right now, the raytracer should work. To play with it, it uses stdout to
output a .ppm image