Add more array functionality
This commit also adds a bench for bubble sort, but, for some god
forsaken reason, it randomly decides to segfault.
You run the same code, 50% of the time it works, the other 50% it
segfaults for some unknown reason.

I think this may signal the time to rewrite the type system, and handle
it properly. As right now, it's a hack on top of hacks. Especially with
that horrible NamedType variant. It's crazy to work with, and most of
the time, things are hacked together.

I guess I'll start rewriting the type system properly, and make things
simpler, and have the types themselves handle read and write operations.
Or at least getting pointers to them properly, being able to send slices
(unsized arrays) to functions and casting them back to the sized
variant, etc. This will help very much with implementing generics, which
are really starting to be needed.

The problems with the type system are becoming obvious with each feature
I try and implement now. The current design is horrible, and it uses
mostly AST type nodes to handle things, there's no concrete system to
help with code generation, safety, generics and things like that.

End of rant.

Update README.md