Add `if ret`, format code, update rust nightly.
Better error reporting is now in place for functions with bodies that
are missing a return value, which can happen from a broken expression
inside them, or an `if` statement instead of `if ret`.

By implementing Deref on ValueTypePair, a lot of the code was refactored
into a nicer form and easier to understand.
As well as that, to fix some issues with if statements and expressions,
I've added `if ret` which turns the `if/else` statement into an
expression.

On top of that, the compiler now uses the latest Rust Nightly for
or_patterns to simplify some pattern matching duplication clauses.